backend/requirements.txt
flask==2.2.5
connexion[swagger-ui]==2.14.2
flask-cors==4.0.0
flask-jwt-extended==4.4.4
flask-sqlalchemy==3.0.3
sqlalchemy==2.0.23
alembic==1.13.0
bcrypt==4.1.1
python-dotenv==1.0.0
